backend/requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Cors==5.0.0
flask-mongoengine==1.0.0
Flask-WTF==1.2.1
frozendict==2.4.4
google-auth==2.35.0
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
mongoengine==0.29.1
multitasking==0.0.11
numpy==2.1.1
pandas==2.2.3
peewee==3.17.6
platformdirs==4.3.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rsa==4.9
six==1.16.0
soupsieve==2.6
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
Werkzeug==3.0.4
WTForms==3.1.2
